// src/types/user.ts
// types/user.ts
export type User = {
    id: string;
    email: string;
    displayName: string | null;
    firstName: string | null;
    lastName: string | null;
    phone: string | null;
    address: string | null;
    city: string | null;
    postalCode: string | null;
    avatarUrl: string | null;
    isPro: boolean | null;
    createdAt: string | null;
};